fix: First commit for making robot agent more modular
