+++ robot_agent/robot_agent/olds/spot_controller.py
@@ -1,8 +1,18 @@
-from rosa import RobotSystemPrompts
+import rclpy.logging
+
+import os
+from dotenv import load_dotenv
+from langchain.tools import Tool
+
+from langchain_openai import ChatOpenAI
+from langchain_ollama import ChatOllama
+from rosa import ROSA, RobotSystemPrompts
 from langchain.agents import tool
 
+import rclpy
 from rclpy.node import Node
 import threading
+from rclpy.executors import MultiThreadedExecutor
 
 from std_msgs.msg import String, Bool
 from geometry_msgs.msg import Twist
@@ -20,22 +30,40 @@ from spot_msgs.msg import (
 from geometry_msgs.msg import TwistWithCovarianceStamped
 from sensor_msgs.msg import JointState
 
-
+import asyncio
 import time
 from datetime import datetime
 
 from typing import List, Optional
 from colorama import Fore, Style, init
+import logging
 
 import json
 
 
 init(autoreset=True)
 
+# Create a logger object named 'log'
+log = logging.getLogger("InteractionLogger")
+log.setLevel(logging.INFO)  # Set the minimum level of messages to record
+
+# Create a file handler to write logs to a file named 'interaction_log.txt'
+file_handler = logging.FileHandler("interaction_log.txt")
+
+# Create a formatter to define the log message format (timestamp - level - message)
+formatter = logging.Formatter(
+    "%(asctime)s - %(levelname)s - %(message)s", datefmt="%Y-%m-%d %H:%M:%S"
+)
+file_handler.setFormatter(formatter)
+
+# Add the file handler to the logger
+log.addHandler(file_handler)
+
 
 tracking_confirmation_received = threading.Event()
 last_published_data = None
 
+load_dotenv()  # This loads the variables from .env file
 
 current_tracking_object = None
 
@@ -95,6 +123,7 @@ class SpotController(Node):
 
         # Publishers
         self.commands_pub = None
+        self.llm_response_pub = None
         self.key_pressed_pub = None
         self.tracking_signal_pub = None
         self.switch_mode_pub = None
@@ -111,6 +140,8 @@ class SpotController(Node):
         self.wifi_state_sub = None
         self.joint_states_sub = None
 
+        self.user_query_sub = None
+
         self.person_info_sub = None
         self.tracking_status_sub = None
 
@@ -122,6 +153,8 @@ class SpotController(Node):
 
         # variables
         self.warned_low_battery = False
+
+        self.user_query = None
 
         self.battery_states = None
         self.metrics = None
@@ -158,6 +191,8 @@ class SpotController(Node):
         self.declare_parameter("wifi_state_topic", self.wifi_state_topic)
         self.declare_parameter("joint_states_topic", self.joint_states_topic)
 
+        self.declare_parameter("user_query_topic", self.user_query_topic)
+        self.declare_parameter("llm_response_topic", self.llm_response_topic)
         self.declare_parameter("key_pressed_topic", self.key_pressed_topic)
         self.declare_parameter("tracking_signal_topic", self.tracking_signal_topic)
         self.declare_parameter("person_info_topic", self.person_info_topic)
@@ -202,6 +237,14 @@ class SpotController(Node):
         )
         self.joint_states_topic = (
             self.get_parameter("joint_states_topic").get_parameter_value().string_value
+        )
+
+        self.user_query_topic = (
+            self.get_parameter("user_query_topic").get_parameter_value().string_value
+        )
+
+        self.llm_response_topic = (
+            self.get_parameter("llm_response_topic").get_parameter_value().string_value
         )
 
         self.key_pressed_topic = (
@@ -366,6 +409,11 @@ class SpotController(Node):
         self.joint_states = msg
         self.get_logger().debug(f"Joint states received: {msg}")
 
+    def user_query_callback(self, msg) -> None:
+        """Callback for the subscriber node (to topic self.user_query_topic).
+        Receives user query messages."""
+        self.user_query = msg.data
+
     def person_info_callback(self, msg) -> None:
         """Callback for the subscriber node (to topic self.person_info_topic).
         Receives information about detected persons and objects.
@@ -464,6 +512,10 @@ class SpotController(Node):
 
         This should be called when necessary to guarantee accurate battery level information.
         """
+
+        time_str = datetime.now().strftime("%H:%M:%S:%f")
+        print(Fore.CYAN + f"[{time_str}] Start battery states time..")
+
         battery_percentage_str: str = "unknown"
         remaining_runtime_str: str = "unknown"
         temperatures_str: str = "unknown"
@@ -480,6 +532,9 @@ class SpotController(Node):
                 f"The current status of the battery is {self.battery_states.status}.\n"
             )
 
+            time_str = datetime.now().strftime("%H:%M:%S:%f")
+            print(Fore.CYAN + f"[{time_str}] End battery states time..")
+
             return (
                 battery_percentage_str
                 + remaining_runtime_str
@@ -488,6 +543,8 @@ class SpotController(Node):
             )
 
         else:
+            time_str = datetime.now().strftime("%H:%M:%S:%f")
+            print(Fore.CYAN + f"[{time_str}] End battery states time..")
             return f"Battery percentage for {ROBOT_NAME} has not been reported yet or is unavailable."
 
     def get_mobility_metrics(self) -> str:
@@ -499,6 +556,10 @@ class SpotController(Node):
         * The current physical posture of the robot, namely whether or not the robot dog is standing/sitting and the desired body position and orientation (pose) relative to the world.
         * The robot’s current motion configuration, namely whether or not the robot is currently moving, a locomotion behavior hint (e.g. walking, stair climbing), and whether the robot is actively in a stair-climbing mode.
         """
+
+        time_str = datetime.now().strftime("%H:%M:%S:%f")
+        print(Fore.CYAN + f"[{time_str}] Start metrics time..")
+
         total_distance_str: str = "unknown"
         total_time_moving: str = "unknown"
         physical_posture: str = "unknown"
@@ -525,6 +586,9 @@ class SpotController(Node):
             physical_posture += f"The estimated body position and orientation (pose) relative to the world is {self.mobility_params.body_control}\n"
             motion_configuration += f"Locomotion behavior hint: {self.mobility_params.locomotion_hint}\nStair climbing: {self.mobility_params.stair_hint}.\n"
 
+        time_str = datetime.now().strftime("%H:%M:%S:%f")
+        print(Fore.CYAN + f"[{time_str}] End metrics time..")
+
         return (
             "Total distance result: "
             + total_distance_str
@@ -545,14 +609,32 @@ class SpotController(Node):
           * the estimated runtime remaining for locomotion.
           * the accumulated electrical power usage.
         """
+        time_str = datetime.now().strftime("%H:%M:%S:%f")
+        print(Fore.CYAN + f"[{time_str}] Start power state time..")
+
+        time_str = datetime.now().strftime("%H:%M:%S:%f")
+        print(Fore.CYAN + f"[{time_str}] End power state time..")
         return "Tool to be implemented."
 
     def get_info_spot(self) -> str:
         """Gets the robot dog (spot) general information, such as its specie, version and nickname"""
+
+        time_str = datetime.now().strftime("%H:%M:%S:%f")
+        print(Fore.CYAN + f"[{time_str}] Start info spot time..")
+
+        time_str = datetime.now().strftime("%H:%M:%S:%f")
+        print(Fore.CYAN + f"[{time_str}] End inof spot time..")
+
         return "Tool to be implemented."
 
     def get_odometry(self) -> str:
         """Represents the robot's current estimated velocity, both linear and angular, in three dimensions."""
+
+        time_str = datetime.now().strftime("%H:%M:%S:%f")
+        print(Fore.CYAN + f"[{time_str}] Start odometry time..")
+
+        time_str = datetime.now().strftime("%H:%M:%S:%f")
+        print(Fore.CYAN + f"[{time_str}] End odometry time..")
         return "Tool to be implemented."
 
     def get_faults(self) -> str:
@@ -563,6 +645,11 @@ class SpotController(Node):
         * System Fault State:
             Reports both active and historical faults detected in the robot’s internal systems.
         """
+        time_str = datetime.now().strftime("%H:%M:%S:%f")
+        print(Fore.CYAN + f"[{time_str}] Start faults time..")
+
+        time_str = datetime.now().strftime("%H:%M:%S:%f")
+        print(Fore.CYAN + f"[{time_str}] End faults time..")
         return "Tool to be implemented."
 
     def get_wifi_connection_state(self) -> str:
@@ -578,6 +665,10 @@ class SpotController(Node):
         Please make sure to communicate the meanings (unknown, access point, client) to the user, as they might not know what a raw number  mean.
 
         """
+
+        time_str = datetime.now().strftime("%H:%M:%S:%f")
+        print(Fore.CYAN + f"[{time_str}] Start wifi time..")
+
         if self.wifi_state is not None:
             mode_str = (
                 "unknown"
@@ -589,8 +680,16 @@ class SpotController(Node):
                 if self.wifi_state.ssid
                 else "unknown"
             )
+
+            time_str = datetime.now().strftime("%H:%M:%S:%f")
+            print(Fore.CYAN + f"[{time_str}] End wifi time..")
+
             return f"Wi-Fi Mode: {mode_str}.\n SSID: {ssid_str}."
         else:
+
+            time_str = datetime.now().strftime("%H:%M:%S:%f")
+            print(Fore.CYAN + f"[{time_str}] End wifi time..")
+
             return "Wi-Fi state has not been reported yet or is unavailable."
 
     def get_general_status(self) -> str:
@@ -603,6 +702,9 @@ class SpotController(Node):
             * get_mobility_metrics for mobility metrics,
             * get_wifi_connection_state for the connection status
         """
+
+        time_str = datetime.now().strftime("%H:%M:%S:%f")
+        print(Fore.CYAN + f"[{time_str}] Start general status time..")
 
         battery_percentage_str: str = "unknown"
         posture_str: str = "unknown"
@@ -627,6 +729,9 @@ class SpotController(Node):
             if current_tracking_object
             else " Not currently tracking."
         )
+
+        time_str = datetime.now().strftime("%H:%M:%S:%f")
+        print(Fore.CYAN + f"[{time_str}] End general status time..")
 
         return (
             "Battery info :"
@@ -662,9 +767,18 @@ class SpotController(Node):
                 print(Fore.CYAN + f"[{time_str}] End sit time..")
 
                 return f"{ROBOT_NAME} is sitting."
+
             except Exception as e:
+                # time stamp
+                time_str = datetime.now().strftime("%H:%M:%S:%f")
+                print(Fore.CYAN + f"[{time_str}] End sit time..")
+
                 return f"Failed to sit {ROBOT_NAME}: {e}"
         else:
+            # time stamp
+            time_str = datetime.now().strftime("%H:%M:%S:%f")
+            print(Fore.CYAN + f"[{time_str}] End sit time..")
+
             return f"Robot state (sitting/standing) is not yet known. Cannot sit."
 
     def stand(self) -> str:
@@ -692,8 +806,15 @@ class SpotController(Node):
 
                 return f"{ROBOT_NAME} is standing."
             except Exception as e:
+
+                time_str = datetime.now().strftime("%H:%M:%S:%f")
+                print(Fore.CYAN + f"[{time_str}] End stand tool time..")
+
                 return f"Failed to stand {ROBOT_NAME}: {e}"
         else:
+            time_str = datetime.now().strftime("%H:%M:%S:%f")
+            print(Fore.CYAN + f"[{time_str}] End stand tool time..")
+
             return f"Robot state (sitting/standing) is not yet known. Cannot stand."
 
     def move(self, linear: List[float], angular: float, duration: int) -> str:
@@ -712,12 +833,13 @@ class SpotController(Node):
         If the user does not specify a time, assume a default duration of 1 second.
 
         :param linear: A list of 3 floats representing [x, y, z] velocity in m/s. This vector should be constructed based on the user's direction and specified speed.
-                    For example, if the user says "go right at 1.2 m/s", the vector should be (0.0, -1.2, 0.0).
+                    For example, if the user says "go right at 1.2 m/s", the vector should be [0.0, -1.2, 0.0].
         :param angular: A float for z-axis angular velocity (rotation).
         :param duration: Duration of the movement in seconds.
 
         Do not attempt to move the robot if the battery level is below 20% or if the robot is sitting. In such cases, return an appropriate message.
         To have the battery level and whether or not the robot is sitting, you have the , use the get_general_status() tool.
+        Remember, for the parameters, you have to provide floats, not integers!
         """
 
         # time stamp
@@ -747,7 +869,7 @@ class SpotController(Node):
                 t0 = self.get_clock().now().nanoseconds
 
                 while (self.get_clock().now().nanoseconds - t0) / 1e9 <= duration:
-                    print((self.get_clock().now().nanoseconds - t0) / 1e9)
+                    # print((self.get_clock().now().nanoseconds - t0) / 1e9)
 
                     self.commands_pub.publish(msg_twist)
 
@@ -760,9 +882,16 @@ class SpotController(Node):
 
                 return f"Moved {ROBOT_NAME} with linear={linear}, angular={angular} for {duration}s and then stopped."
             except Exception as e:
+                # time stamp
+                time_str = datetime.now().strftime("%H:%M:%S:%f")
+                print(Fore.CYAN + f"[{time_str}] End motion time..")
+
                 print(f"Failed to move {ROBOT_NAME}: {e}")
                 return f"Failed to move {ROBOT_NAME}: {e}"
         else:
+            # time stamp
+            time_str = datetime.now().strftime("%H:%M:%S:%f")
+            print(Fore.CYAN + f"[{time_str}] End motion time..")
             return f"Robot state (sitting/standing) is not yet known. Cannot move."
 
     def switch_mode(self, mode: str, object_name: Optional[str] = None) -> str:
@@ -805,6 +934,10 @@ class SpotController(Node):
 
             elif mode_requested == "tracking":
                 if not object_name:
+                    # time stamp
+                    time_str_end = datetime.now().strftime("%H:%M:%S:%f")
+                    print(Fore.CYAN + f"[{time_str_end}] End switch mode time..")
+
                     return "Error: To switch to tracking mode, you must specify an object_name."
                 msg_str.data = "t"
                 self.key_pressed_pub.publish(msg_str)
@@ -826,6 +959,10 @@ class SpotController(Node):
             return response
 
         except Exception as e:
+            # time stamp
+            time_str_end = datetime.now().strftime("%H:%M:%S:%f")
+            print(Fore.CYAN + f"[{time_str_end}] End switch mode time..")
+
             return (
                 f"Failed to switch mode for {ROBOT_NAME} to '{mode_requested}' "
                 f"(as '{msg_str.data}'): {e}"
@@ -916,207 +1053,361 @@ class SpotController(Node):
         else:
             return f"A stop command was sent to {ROBOT_NAME} to ensure tracking is disabled."
 
-    ############################################# Getters #########################################################################
-    def get_prompts():
-        prompts = RobotSystemPrompts(
-            embodiment_and_persona="You are a robotic agent managing a robot dog. The robot dog you are operating is Boston dynamics' Spot",
-            about_your_capabilities="You capabilities are limited to the available tools. Anything that is asked to you and not provided by a tool is beyond your capabilities",
-            critical_instructions="Always use the corresponding tool if you can. If the user ask you to perform an action requiring to move the robot, always use the mode tool. Sam for all other tools: if the user ask for an information/action requiring to use a tool, always use the relevant tool. Also, tell the user what you are trying to do.",
-        )
-        return prompts
 
-    def get_tools(self):
+# def print_response(query: str):
+#     response = rosa.invoke(query)
+#     # ... code to display response ...
+################################################### Tools #############################################################
 
-        @tool
-        def get_battery_status():
-            """Provides detailed information about each battery installed in the robot.
-            For instance, it includes :
-            * the percentage of charge remaining,
-            * the estimated remaining runtime,
-            * the temperature readings.
-            * a status indicator to flag battery health or errors.
 
-            The status indicator is a number with the following meaning:
-                - 0 -> STATUS UNKNOWN
-                - 1 -> STATUS MISSING
-                - 2 -> STATUS CHARGING
-                - 3 -> STATUS DISCHARGING
-                - 4 -> STATUS BOOTING
-            Please make sure to communicate the meanings to the user, as they might not know what a raw number for the battery status mean.
+async def submit(query: str, rosa):
+    return await stream_response(query, rosa)
 
-            This should be called when necessary to guarantee accurate battery level information.
-            """
-            return self.get_battery_status()
 
-        @tool
-        def get_mobility_metrics():
-            """
-            Reports mobility statistics of the robot dog.
-            The information that you can get through this tool include:
-            * The total distance traveled by the robot,
-            * The total time the robot has been moving,
-            * The current physical posture of the robot, namely whether or not the robot dog is standing/sitting and the desired body position and orientation (pose) relative to the world.
-            * The robot’s current motion configuration, namely whether or not the robot is currently moving, a locomotion behavior hint (e.g. walking, stair climbing), and whether the robot is actively in a stair-climbing mode.
-            """
-            return self.get_mobility_metrics()
+async def stream_response(query: str, rosa):
+    print(Fore.BLUE + Style.BRIGHT + f"\n👤 User: {query}\n")
 
-        @tool
-        def get_power_state():
-            """
-            Provides the current power supply conditions.
-            The information that you can get through this tool include:
-            * the state of the motor and shore power systems,
-            * the charge level specific to locomotion systems,
-            * the estimated runtime remaining for locomotion.
-            * the accumulated electrical power usage.
-            """
-            return self.get_power_state()
+    response = ""
 
-        @tool
-        def get_info_spot():
-            """Gets the robot dog (spot) general information, such as its specie, version and nickname"""
-            return self.get_info_spot()
+    async for event in rosa.astream(query):
+        if event["type"] == "token":
+            print(Fore.GREEN + event["content"], end="", flush=True)
+            response = response + event["content"]
+        elif event["type"] == "tool_start":
+            print(Fore.YELLOW + f"\n🛠️ Starting tool: {event['name']}")
+        elif event["type"] == "tool_end":
+            print(Fore.YELLOW + f"\n✅ Finished tool: {event['name']}")
+            await asyncio.sleep(1)
+        elif event["type"] == "final":
+            pass
+            # print(Fore.CYAN + Style.BRIGHT + f"\n📤 Final output: {event['content']}")
+        elif event["type"] == "error":
+            print(Fore.RED + f"\n❌ Error: {event['content']}")
 
-        @tool
-        def get_odometry():
-            """Represents the robot's current estimated velocity, both linear and angular, in three dimensions."""
-            return self.get_odometry()
+    return response
 
-        @tool
-        def get_faults():
-            """
-            Gets faults reports of the robot dog. There are two faults types:
-            * Behavior Fault State:
-                Lists any active behavior-related faults.
-            * System Fault State:
-                Reports both active and historical faults detected in the robot’s internal systems.
-            """
-            return self.get_faults()
 
-        @tool
-        def get_wifi_connection_state():
-            """Conveys the current Wi-Fi connection status of the robot.
-            The information that you can get through this tool include:
-            * the current operating mode of the wireless interface
-            * the name (SSID) of the currently connected network if applicable
-            The operating mode  is a number with the following meaning:
-                - 0 -> MODE UNKNOWN
-                - 1 -> MODE ACCESS POINT
-                - 2 -> MODE CLIENT
+# Function to run the ROSA agent
+async def run(node, rosa):
+    # Print the time when the session starts
+    start_time_str = time.strftime("%Y-%m-%d %H:%M:%S")
+    print(
+        Fore.CYAN + Style.BRIGHT + f"\n--- ROSA Session Started at {start_time_str} ---"
+    )
 
-            Please make sure to communicate the meanings (unknown, access point, client) to the user, as they might not know what a raw number  mean.
+    query = None
 
-            """
-            return self.get_wifi_connection_state()
+    while True:
 
-        @tool
-        def get_general_status():
-            """Gets the current robot dog's latest reported state and general information.
-            The status includes battery information, the robot's posture (standing, sitting, moving), wifi connection status, and the whether or not tracking is ongoing.
-            You should use this tool if the user simply asks for the status of the robot.
-            However, if they ask more details, please use the appropriate tools.
-            For more detailed information, use the appropriate tools, namely
-                * get_battery_status for battery information,
-                * get_mobility_metrics for mobility metrics,
-                * get_wifi_connection_state for the connection status
-            """
-            return self.get_general_status()
+        # # Get the user's command
+        # if node.user_query != query:
+        #     query = node.user_query
 
-        @tool
-        def sit():
-            """Command the robot dog to sit and transition from a standing position (on 4 legs) to a sitting position.
-            It cannot be used if the robot is already on the sitting. To know whether or not the robot is sitting, first get the status of the robot.
-            """
-            return self.sit()
+        #     # --- 1. Print the time BEFORE the prompt ---
+        #     prompt_time_str = time.strftime("%H:%M:%S")
+        #     # The '\n' adds a space before the new prompt, 'end=""' keeps the cursor on the same line
+        #     print(Fore.CYAN + f"\n[{prompt_time_str}] ", end="")
 
-        @tool
-        def stand():
-            """Command the robot to stand and transition from a sitting position to a standing position.
-            It cannot be used if the robot is already standing.  To know whether or not the robot is sitting, first get the status of the robot.
-            Do not attempt to stand the robot if the battery level is below 20% or if the robot is standing. In such cases, return an appropriate message.
-            To have the battery level, use the get_battery_status() tool.
-            """
-            return self.stand()
+        #     query = input("Enter your prompt (or 'exit' to quit): ")
 
-        @tool
-        def move(linear, angular, duration):
-            """
-            Move the robot with specified linear and angular velocities for a given duration.
-            This function controls movement along three axes (x, y, z) and rotation.
+        #     if query.lower() == "exit":
+        #         log.info("--- Session Ended ---")
+        #         break
 
-            The coordinate system is as follows:
-            - x-axis: +x is forward, -x is backward.
-            - y-axis: +y is left, -y is right.
-            - z-axis: +z is up, -z is down.
-            - angular z-axis: +z is counter-clockwise turn (left), -z is clockwise turn (right).
+        #     # Log the prompt to the file
+        #     log.info(f"USER_PROMPT: {query}")
 
-            To perform this movement, the robot dog must be in the standing.
-            The user may specify a speed (e.g., "velocity 1m/s", "go slowly"). If a speed is provided, use it to set the magnitude of the linear or angular velocity vector. If no speed is specified, use a default of 1 m/s or -1 m/s.
-            If the user does not specify a time, assume a default duration of 1 second.
+        #     # Start processing the command
+        #     processing_time_str = time.strftime("%H:%M:%S")
+        #     print(Fore.CYAN + f"[{processing_time_str}] Processing command...")
 
-            :param linear: A list of 3 floats representing [x, y, z] velocity in m/s. This vector should be constructed based on the user's direction and specified speed.
-                        For example, if the user says "go right at 1.2 m/s", the vector should be [0.0, -1.2, 0.0].
-            :param angular: A float for z-axis angular velocity (rotation).
-            :param duration: Duration of the movement in seconds.
+        #     response = await submit(query, rosa)
 
-            Do not attempt to move the robot if the battery level is below 20% or if the robot is sitting. In such cases, return an appropriate message.
-            To have the battery level and whether or not the robot is sitting, you have the , use the get_general_status() tool.
-            """
-            return self.move(linear, angular, duration)
+        #     responseMsg = String()
+        #     responseMsg.data = response
+        #     node.llm_response_pub.publish(responseMsg)
 
-        @tool
-        def switch_mode(mode, object_name):
-            """
-            Switches the control mode of the robot dog. Tell the user that he has to select the image window.
-            The LLM should request modes like 'keyboard', 'hand', 'tracking' or 'stop tracking'.
-            If the user selects 'keyboard', he has to know that to stand he has to use "t", to sit "g", to move the letters "a", "w", "d", "s". Still in keyboard mode
-            the user must know that for rotation he/she can use "left-arrow" and "right-arrow" to move left or right respectively, and for altitude, he/she can use "up-arrow" to move up and "down-arrow" to move down.
-            If the user selects 'hand', he has to know that he has to use the hands to control the robot dog, all the options are in the image window.
-            If the user selects 'tracking', tracking': Start tracking a person holding a specific object.
-            When using this mode, you must also provide the 'object_name' parameter.
-            Choose the object from this list: [backpack, umbrella, handbag, bottle, cup, fork, knife, spoon, bowl, banana, apple, cell phone, book].
-            If the user selects 'stop tracking': Stop the current tracking task.
+        #     processing_time_str = time.strftime("%H:%M:%S")
+        #     print(Fore.CYAN + f"[{processing_time_str}] Response received...")
 
-            :param mode: The desired control mode as a string (e.g., "keyboard", "hand", "tracking", "stop tracking").
-            :param object_name: The name of the object to track. Required only for 'tracking' mode.
-            """
-            return self.switch_mode(mode, object_name)
+        # --- 1. Print the time BEFORE the prompt ---
+        prompt_time_str = time.strftime("%H:%M:%S")
+        # The '\n' adds a space before the new prompt, 'end=""' keeps the cursor on the same line
+        print(Fore.CYAN + f"\n[{prompt_time_str}] ", end="")
 
-        @tool
-        def start_object_tracking(object_name):
-            """
-            Use this tool to start tracking a person holding a specific object. It will wait
-            up to 5 seconds for a person holding this object to be detected. The robot MUST be  standing.
+        query = input("Enter your prompt (or 'exit' to quit): ")
 
-            First, you MUST choose the most similar object from this list of available options:
-            [backpack, umbrella, handbag, bottle, cup, fork, knife, spoon, bowl, banana, apple, cell phone, book]
-            Match the user's request to an object in the list. For example, if the user asks for a "phone", choose "cell phone".
-            If you cannot find a clear match, respond by saying "There are no similar objects to track."
-            :param object_name: str - The chosen object name from the list.
-            """
-            return self.start_object_tracking(object_name)
+        if query.lower() == "exit":
+            log.info("--- Session Ended ---")
+            break
 
-        @tool
-        def stop_object_tracking():
-            """
-            Stops tracking the current object and clears the tracking target.
-            An explicit "stop_tracking" message is sent to the tracking system.
-            """
-            return self.stop_object_tracking()
+        # Log the prompt to the file
+        log.info(f"USER_PROMPT: {query}")
 
-        return [
-            get_battery_status,
-            get_mobility_metrics,
-            get_info_spot,
-            get_faults,
-            get_odometry,
-            get_general_status,
-            get_power_state,
-            stand,
-            sit,
-            start_object_tracking,
-            stop_object_tracking,
-            move,
-            switch_mode,
-            get_wifi_connection_state,
-        ]
+        # Start processing the command
+        processing_time_str = time.strftime("%H:%M:%S")
+        print(Fore.CYAN + f"[{processing_time_str}] Processing command...")
+
+        response = await submit(query, rosa)
+
+        responseMsg = String()
+        responseMsg.data = response
+        node.llm_response_pub.publish(responseMsg)
+
+        processing_time_str = time.strftime("%H:%M:%S")
+        print(Fore.CYAN + f"[{processing_time_str}] Response received...")
+
+
+# asyncio.run(run())
+
+
+# ROS init and run
+def main(args=None):
+    rclpy.init(args=args)
+    node = SpotController()
+    # node.get_logger().set_level(rclpy.logging.LoggingSeverity.DEBUG)
+
+    # Use executor in a separate thread
+    executor = MultiThreadedExecutor()
+    executor.add_node(node)
+
+    def spin():
+        executor.spin()
+
+    # Start the ROS spinning in a background thread
+    spin_thread = threading.Thread(target=spin, daemon=True)
+    spin_thread.start()
+
+    prompts = RobotSystemPrompts(
+        embodiment_and_persona="You are a robotic agent managing a robot dog. The robot dog you are operating is Boston dynamics' Spot",
+        about_your_capabilities="You capabilities are limited to the available tools. Anything that is asked to you and not provided by a tool is beyond your capabilities",
+        critical_instructions="Always use the corresponding tool if you can. If the user ask you to perform an action requiring to move the robot, always use the mode tool. Sam for all other tools: if the user ask for an information/action requiring to use a tool, always use the relevant tool. Also, tell the user what you are trying to do.",
+    )
+
+    ################################## tools ########################################################
+
+    @tool
+    def get_battery_status(node=node):
+        """Provides detailed information about each battery installed in the robot.
+        For instance, it includes :
+        * the percentage of charge remaining,
+        * the estimated remaining runtime,
+        * the temperature readings.
+        * a status indicator to flag battery health or errors.
+
+        The status indicator is a number with the following meaning:
+            - 0 -> STATUS UNKNOWN
+            - 1 -> STATUS MISSING
+            - 2 -> STATUS CHARGING
+            - 3 -> STATUS DISCHARGING
+            - 4 -> STATUS BOOTING
+        Please make sure to communicate the meanings to the user, as they might not know what a raw number for the battery status mean.
+
+        This should be called when necessary to guarantee accurate battery level information.
+        """
+        return node.get_battery_status()
+
+    @tool
+    def get_mobility_metrics(node=node):
+        """
+        Reports mobility statistics of the robot dog.
+        The information that you can get through this tool include:
+        * The total distance traveled by the robot,
+        * The total time the robot has been moving,
+        * The current physical posture of the robot, namely whether or not the robot dog is standing/sitting and the desired body position and orientation (pose) relative to the world.
+        * The robot’s current motion configuration, namely whether or not the robot is currently moving, a locomotion behavior hint (e.g. walking, stair climbing), and whether the robot is actively in a stair-climbing mode.
+        """
+        return node.get_mobility_metrics()
+
+    @tool
+    def get_power_state(node=node):
+        """
+        Provides the current power supply conditions.
+        The information that you can get through this tool include:
+        * the state of the motor and shore power systems,
+        * the charge level specific to locomotion systems,
+        * the estimated runtime remaining for locomotion.
+        * the accumulated electrical power usage.
+        """
+        return node.get_power_state()
+
+    @tool
+    def get_info_spot(node=node):
+        """Gets the robot dog (spot) general information, such as its specie, version and nickname"""
+        return node.get_info_spot()
+
+    @tool
+    def get_odometry(node=node):
+        """Represents the robot's current estimated velocity, both linear and angular, in three dimensions."""
+        return node.get_odometry()
+
+    @tool
+    def get_faults(node=node):
+        """
+        Gets faults reports of the robot dog. There are two faults types:
+        * Behavior Fault State:
+            Lists any active behavior-related faults.
+        * System Fault State:
+            Reports both active and historical faults detected in the robot’s internal systems.
+        """
+        return node.get_faults()
+
+    @tool
+    def get_wifi_connection_state(node=node):
+        """Conveys the current Wi-Fi connection status of the robot.
+        The information that you can get through this tool include:
+        * the current operating mode of the wireless interface
+        * the name (SSID) of the currently connected network if applicable
+        The operating mode  is a number with the following meaning:
+            - 0 -> MODE UNKNOWN
+            - 1 -> MODE ACCESS POINT
+            - 2 -> MODE CLIENT
+
+        Please make sure to communicate the meanings (unknown, access point, client) to the user, as they might not know what a raw number  mean.
+
+        """
+        return node.get_wifi_connection_state()
+
+    @tool
+    def get_general_status(node=node):
+        """Gets the current robot dog's latest reported state and general information.
+        The status includes battery information, the robot's posture (standing, sitting, moving), wifi connection status, and the whether or not tracking is ongoing.
+        You should use this tool if the user simply asks for the status of the robot.
+        However, if they ask more details, please use the appropriate tools.
+        For more detailed information, use the appropriate tools, namely
+            * get_battery_status for battery information,
+            * get_mobility_metrics for mobility metrics,
+            * get_wifi_connection_state for the connection status
+        """
+        return node.get_general_status()
+
+    @tool
+    def sit(node=node):
+        """Command the robot dog to sit and transition from a standing position (on 4 legs) to a sitting position.
+        It cannot be used if the robot is already on the sitting. To know whether or not the robot is sitting, first get the status of the robot.
+        """
+        return node.sit()
+
+    @tool
+    def stand(node=node):
+        """Command the robot to stand and transition from a sitting position to a standing position.
+        It cannot be used if the robot is already standing.  To know whether or not the robot is sitting, first get the status of the robot.
+        Do not attempt to stand the robot if the battery level is below 20% or if the robot is standing. In such cases, return an appropriate message.
+        To have the battery level, use the get_battery_status() tool.
+        """
+        return node.stand()
+
+    @tool
+    def move(linear, angular, duration, node=node):
+        """
+        Move the robot with specified linear and angular velocities for a given duration.
+        This function controls movement along three axes (x, y, z) and rotation.
+
+        The coordinate system is as follows:
+        - x-axis: +x is forward, -x is backward.
+        - y-axis: +y is left, -y is right.
+        - z-axis: +z is up, -z is down.
+        - angular z-axis: +z is counter-clockwise turn (left), -z is clockwise turn (right).
+
+        To perform this movement, the robot dog must be in the standing.
+        The user may specify a speed (e.g., "velocity 1m/s", "go slowly"). If a speed is provided, use it to set the magnitude of the linear or angular velocity vector. If no speed is specified, use a default of 1 m/s or -1 m/s.
+        If the user does not specify a time, assume a default duration of 1 second.
+
+        :param linear: A list of 3 floats representing [x, y, z] velocity in m/s. This vector should be constructed based on the user's direction and specified speed.
+                    For example, if the user says "go right at 1.2 m/s", the vector should be [0.0, -1.2, 0.0].
+        :param angular: A float for z-axis angular velocity (rotation).
+        :param duration: Duration of the movement in seconds.
+
+        Do not attempt to move the robot if the battery level is below 20% or if the robot is sitting. In such cases, return an appropriate message.
+        To have the battery level and whether or not the robot is sitting, you have the , use the get_general_status() tool.
+        Remember, for the parameters, you have to provide floats, not integers!
+        """
+        return node.move(linear, angular, duration)
+
+    @tool
+    def switch_mode(mode, object_name, node=node):
+        """
+        Switches the control mode of the robot dog. Tell the user that he has to select the image window.
+        The LLM should request modes like 'keyboard', 'hand', 'tracking' or 'stop tracking'.
+        If the user selects 'keyboard', he has to know that to stand he has to use "t", to sit "g", to move the letters "a", "w", "d", "s". Still in keyboard mode
+        the user must know that for rotation he/she can use "left-arrow" and "right-arrow" to move left or right respectively, and for altitude, he/she can use "up-arrow" to move up and "down-arrow" to move down.
+        If the user selects 'hand', he has to know that he has to use the hands to control the robot dog, all the options are in the image window.
+        If the user selects 'tracking', tracking': Start tracking a person holding a specific object.
+        When using this mode, you must also provide the 'object_name' parameter.
+        Choose the object from this list: [backpack, umbrella, handbag, bottle, cup, fork, knife, spoon, bowl, banana, apple, cell phone, book].
+        If the user selects 'stop tracking': Stop the current tracking task.
+
+        :param mode: The desired control mode as a string (e.g., "keyboard", "hand", "tracking", "stop tracking").
+        :param object_name: The name of the object to track. Required only for 'tracking' mode.
+        """
+        return node.switch_mode(mode, object_name)
+
+    @tool
+    def start_object_tracking(object_name, node=node):
+        """
+        Use this tool to start tracking a person holding a specific object. It will wait
+        up to 5 seconds for a person holding this object to be detected. The robot MUST be  standing.
+
+        First, you MUST choose the most similar object from this list of available options:
+        [backpack, umbrella, handbag, bottle, cup, fork, knife, spoon, bowl, banana, apple, cell phone, book]
+        Match the user's request to an object in the list. For example, if the user asks for a "phone", choose "cell phone".
+        If you cannot find a clear match, respond by saying "There are no similar objects to track."
+        :param object_name: str - The chosen object name from the list.
+        """
+        return node.start_object_tracking(object_name)
+
+    @tool
+    def stop_object_tracking(node=node):
+        """
+        Stops tracking the current object and clears the tracking target.
+        An explicit "stop_tracking" message is sent to the tracking system.
+        """
+        return node.stop_object_tracking()
+
+    tools = [
+        get_battery_status,
+        get_mobility_metrics,
+        get_info_spot,
+        get_faults,
+        get_odometry,
+        get_general_status,
+        get_power_state,
+        stand,
+        sit,
+        start_object_tracking,
+        stop_object_tracking,
+        move,
+        switch_mode,
+        get_wifi_connection_state,
+    ]
+
+    llm_model = ChatOpenAI(
+        model="gpt-4-turbo",
+        temperature=0,
+        timeout=None,
+        max_retries=2,
+    )
+
+    # Pass the LLM to ROSA
+    rosa = ROSA(
+        ros_version=2, llm=llm_model, streaming=True, tools=tools, prompts=prompts
+    )
+
+    try:
+        asyncio.run(run(node, rosa))
+    except KeyboardInterrupt:
+        pass
+    finally:
+        executor.shutdown()
+        node.destroy_node()
+        rclpy.shutdown()
+
+
+if __name__ == "__main__":
+    main()
+
+
+#     def tracking_signal_print_callback(self, msg: String):
+#         self.get_logger().info("--- FINAL TRACKING SIGNAL (Output) ---")
+#         self.get_logger().info(f"{msg.data}")
+#         self.get_logger().info("------------------------------------")
