+++ robot_agent/robot_agent/spot_controller.py
@@ -1,18 +1,7 @@
-import rclpy.logging
-
-import os
-from dotenv import load_dotenv
-from langchain.tools import Tool
-
-from langchain_openai import ChatOpenAI
-from langchain_ollama import ChatOllama
-from rosa import ROSA, RobotSystemPrompts
+from rosa import RobotSystemPrompts
 from langchain.agents import tool
 
-import rclpy
 from rclpy.node import Node
-import threading
-from rclpy.executors import MultiThreadedExecutor
 
 from std_msgs.msg import String, Bool
 from geometry_msgs.msg import Twist
@@ -30,42 +19,15 @@ from spot_msgs.msg import (
 from geometry_msgs.msg import TwistWithCovarianceStamped
 from sensor_msgs.msg import JointState
 
-import asyncio
 import time
 from datetime import datetime
+import threading
 
 from typing import List, Optional
-from colorama import Fore, Style, init
-import logging
-
+from colorama import Fore, init
 import json
 
-
 init(autoreset=True)
-
-# Create a logger object named 'log'
-log = logging.getLogger("InteractionLogger")
-log.setLevel(logging.INFO)  # Set the minimum level of messages to record
-
-# Create a file handler to write logs to a file named 'interaction_log.txt'
-file_handler = logging.FileHandler("interaction_log.txt")
-
-# Create a formatter to define the log message format (timestamp - level - message)
-formatter = logging.Formatter(
-    "%(asctime)s - %(levelname)s - %(message)s", datefmt="%Y-%m-%d %H:%M:%S"
-)
-file_handler.setFormatter(formatter)
-
-# Add the file handler to the logger
-log.addHandler(file_handler)
-
-
-tracking_confirmation_received = threading.Event()
-last_published_data = None
-
-load_dotenv()  # This loads the variables from .env file
-
-current_tracking_object = None
 
 
 DICTIONARY_YOLO_OBJECTS = {
@@ -84,14 +46,18 @@ DICTIONARY_YOLO_OBJECTS = {
     73: "book",
 }
 
-ROBOT_NAME = "Spot"
-
 
 class SpotController(Node):
 
-    # Topics
+    ### Topics and services
+    # velocity commands
     commands_topic = "/byte/cmd_vel"
 
+    # mode switching
+    key_pressed_topic = "/key_pressed"
+    switch_mode_topic = "/switch_mode"
+
+    # Status topics
     battery_state_topic = "/byte/status/battery_states"
     metrics_topic = "/byte/status/metrics"
     feedback_topic = "/byte/status/feedback"
@@ -103,30 +69,23 @@ class SpotController(Node):
     wifi_state_topic = "/byte/status/wifi"
     joint_states_topic = "/byte/joint_states"
 
-    user_query_topic = "/user_query"
-    llm_response_topic = "/llm_response"
-
-    key_pressed_topic = "/key_pressed"
-    switch_mode_topic = "/switch_mode"
-
-    # Topic on which th signal to track a specific person is sent
-    tracking_signal_topic = "/tracking_signal"
-    person_info_topic = "/tracking_info"  # topic on which information about detected persons and objects are sent
-    tracking_status_topic = "/tracking_status"  # topic specifying whether or not the tracking is still on going
-
     # Services
     stand_service_name = "/byte/stand"
     sit_service_name = "/byte/sit"
 
-    def __init__(self):
+    # tracking topics
+    tracking_status_topic = "/tracking_status"  # Topic specifying whether or not the tracking is still ongoing
+    tracking_signal_topic = "/tracking_signal"  # Topic on which th signal to track a specific person is sent
+    person_info_topic = "/tracking_info"  # Topic on which information about detected persons and objects are sent
+
+    def __init__(self, robot_name: str = "Spot") -> None:
         super().__init__("spot_controller")
 
         # Publishers
         self.commands_pub = None
-        self.llm_response_pub = None
         self.key_pressed_pub = None
-        self.tracking_signal_pub = None
         self.switch_mode_pub = None
+        self.tracking_signal_pub = None
 
         # subscribers
         self.battery_state_sub = None
@@ -139,22 +98,15 @@ class SpotController(Node):
         self.system_faults_sub = None
         self.wifi_state_sub = None
         self.joint_states_sub = None
-
-        self.user_query_sub = None
-
         self.person_info_sub = None
         self.tracking_status_sub = None
 
-        # New Subscribers
-
         # clients
-        self.stand_client = None
+        self.stand_clientnode = None
         self.sit_client = None
 
         # variables
         self.warned_low_battery = False
-
-        self.user_query = None
 
         self.battery_states = None
         self.metrics = None
@@ -167,7 +119,11 @@ class SpotController(Node):
         self.wifi_state = None
         self.joint_states = None
 
+        self.robot_name = robot_name
+        self.current_tracking_object = None
         self.tracking_status = None
+
+        self.tracking_confirmation_received = threading.Event()
 
         # Initialize ROS node
         self._init_parameters()
@@ -179,6 +135,8 @@ class SpotController(Node):
     def _init_parameters(self) -> None:
         """Method to initialize parameters such as ROS topics' names"""
         self.declare_parameter("commands_topic", self.commands_topic)
+        self.declare_parameter("key_pressed_topic", self.key_pressed_topic)
+        self.declare_parameter("switch_mode_topic", self.switch_mode_topic)
 
         self.declare_parameter("battery_state_topic", self.battery_state_topic)
         self.declare_parameter("metrics_topic", self.metrics_topic)
@@ -191,19 +149,22 @@ class SpotController(Node):
         self.declare_parameter("wifi_state_topic", self.wifi_state_topic)
         self.declare_parameter("joint_states_topic", self.joint_states_topic)
 
-        self.declare_parameter("user_query_topic", self.user_query_topic)
-        self.declare_parameter("llm_response_topic", self.llm_response_topic)
-        self.declare_parameter("key_pressed_topic", self.key_pressed_topic)
-        self.declare_parameter("tracking_signal_topic", self.tracking_signal_topic)
-        self.declare_parameter("person_info_topic", self.person_info_topic)
-        self.declare_parameter("tracking_status_topic", self.tracking_status_topic)
         self.declare_parameter("stand_service_name", self.stand_service_name)
         self.declare_parameter("sit_service_name", self.sit_service_name)
 
-        self.declare_parameter("switch_mode_topic", self.switch_mode_topic)
+        self.declare_parameter("tracking_status_topic", self.tracking_status_topic)
+        self.declare_parameter("tracking_signal_topic", self.tracking_signal_topic)
+        self.declare_parameter("person_info_topic", self.person_info_topic)
 
         self.commands_topic = (
             self.get_parameter("commands_topic").get_parameter_value().string_value
+        )
+
+        self.key_pressed_topic = (
+            self.get_parameter("key_pressed_topic").get_parameter_value().string_value
+        )
+        self.switch_mode_topic = (
+            self.get_parameter("switch_mode_topic").get_parameter_value().string_value
         )
 
         self.battery_state_topic = (
@@ -239,16 +200,18 @@ class SpotController(Node):
             self.get_parameter("joint_states_topic").get_parameter_value().string_value
         )
 
-        self.user_query_topic = (
-            self.get_parameter("user_query_topic").get_parameter_value().string_value
+        self.stand_service_name = (
+            self.get_parameter("stand_service_name").get_parameter_value().string_value
         )
 
-        self.llm_response_topic = (
-            self.get_parameter("llm_response_topic").get_parameter_value().string_value
+        self.sit_service_name = (
+            self.get_parameter("sit_service_name").get_parameter_value().string_value
         )
 
-        self.key_pressed_topic = (
-            self.get_parameter("key_pressed_topic").get_parameter_value().string_value
+        self.tracking_status_topic = (
+            self.get_parameter("tracking_status_topic")
+            .get_parameter_value()
+            .string_value
         )
 
         self.tracking_signal_topic = (
@@ -261,36 +224,14 @@ class SpotController(Node):
             self.get_parameter("person_info_topic").get_parameter_value().string_value
         )
 
-        self.tracking_status_topic = (
-            self.get_parameter("tracking_status_topic")
-            .get_parameter_value()
-            .string_value
-        )
-
-        self.stand_service_name = (
-            self.get_parameter("stand_service_name").get_parameter_value().string_value
-        )
-
-        self.sit_service_name = (
-            self.get_parameter("sit_service_name").get_parameter_value().string_value
-        )
-
-        self.switch_mode_topic = (
-            self.get_parameter("switch_mode_topic").get_parameter_value().string_value
-        )
-
     def _init_publishers(self) -> None:
         """Method to initialize publishers"""
         self.commands_pub = self.create_publisher(Twist, self.commands_topic, 10)
-        self.llm_response_pub = self.create_publisher(
-            String, self.llm_response_topic, 10
-        )
         self.key_pressed_pub = self.create_publisher(String, self.key_pressed_topic, 10)
+        self.switch_mode_pub = self.create_publisher(String, self.switch_mode_topic, 10)
         self.tracking_signal_pub = self.create_publisher(
             String, self.tracking_signal_topic, 10
         )
-
-        self.switch_mode_pub = self.create_publisher(String, self.switch_mode_topic, 10)
 
     def _init_subscriptions(self) -> None:
         """Method to initialize subscriptions"""
@@ -337,15 +278,12 @@ class SpotController(Node):
             JointState, self.joint_states_topic, self.joint_states_callback, 10
         )
 
-        self.user_query_sub = self.create_subscription(
-            String, self.user_query_topic, self.user_query_callback, 10
+        self.tracking_status_sub = self.create_subscription(
+            Bool, self.tracking_status_topic, self.tracking_status_callback, 10
         )
 
         self.person_info_sub = self.create_subscription(
             String, self.person_info_topic, self.person_info_callback, 10
-        )
-        self.tracking_status_sub = self.create_subscription(
-            Bool, self.tracking_status_topic, self.tracking_status_callback, 10
         )
 
     def _init_clients(self) -> None:
@@ -365,7 +303,7 @@ class SpotController(Node):
             if battery_percentage < 20.0 and not self.warned_low_battery:
                 self.warned_low_battery = True
                 self.get_logger().warn(
-                    f"Battery for {ROBOT_NAME} is at {battery_percentage:.1f}%. "
+                    f"Battery for {self.robot_name} is at {battery_percentage:.1f}%. "
                     "Robot should sit."
                 )
             elif battery_percentage >= 20.0:
@@ -409,21 +347,15 @@ class SpotController(Node):
         self.joint_states = msg
         self.get_logger().debug(f"Joint states received: {msg}")
 
-    def user_query_callback(self, msg) -> None:
-        """Callback for the subscriber node (to topic self.user_query_topic).
-        Receives user query messages."""
-        self.user_query = msg.data
-
     def person_info_callback(self, msg) -> None:
         """Callback for the subscriber node (to topic self.person_info_topic).
         Receives information about detected persons and objects.
         This callback finds a person who is associated with the target object,
         publishes their data ONCE to /tracking_signal, and then stops until a new command.
         """
-        global current_tracking_object, tracking_confirmation_received, last_published_data
 
         # 1. If we are not actively looking for an object, do nothing.
-        if not current_tracking_object:
+        if not self.current_tracking_object:
             return
         if not msg.data or not msg.data.strip():
             return
@@ -443,7 +375,7 @@ class SpotController(Node):
                     held_objects = person_data["info"]["objects"]
 
                     # If the target object is found with this person...
-                    if current_tracking_object in held_objects:
+                    if self.current_tracking_object in held_objects:
 
                         # 4. We found the compatible person! Publish their complete data.
                         self.get_logger().info(
@@ -452,10 +384,8 @@ class SpotController(Node):
 
                         filtered_info = json.dumps(person_data)
                         self.tracking_signal_pub.publish(String(data=filtered_info))
-
-                        last_published_data = filtered_info
-                        tracking_confirmation_received.set()
-                        current_tracking_object = None
+                        self.current_tracking_object = None
+                        self.tracking_confirmation_received.set()
 
                         break  # Stop checking other people in this message
 
@@ -495,27 +425,6 @@ class SpotController(Node):
 
     ################################################## Tools for LLM Commands ########################################################
     def get_battery_status(self) -> str:
-        """Provides detailed information about each battery installed in the robot.
-        For instance, it includes :
-         * the percentage of charge remaining,
-         * the estimated remaining runtime,
-         * the temperature readings.
-         * a status indicator to flag battery health or errors.
-
-         The status indicator is a number with the following meaning:
-            - 0 -> STATUS UNKNOWN
-            - 1 -> STATUS MISSING
-            - 2 -> STATUS CHARGING
-            - 3 -> STATUS DISCHARGING
-            - 4 -> STATUS BOOTING
-        Please make sure to communicate the meanings to the user, as they might not know what a raw number for the battery status mean.
-
-        This should be called when necessary to guarantee accurate battery level information.
-        """
-
-        time_str = datetime.now().strftime("%H:%M:%S:%f")
-        print(Fore.CYAN + f"[{time_str}] Start battery states time..")
-
         battery_percentage_str: str = "unknown"
         remaining_runtime_str: str = "unknown"
         temperatures_str: str = "unknown"
@@ -523,7 +432,7 @@ class SpotController(Node):
 
         if self.battery_states is not None:
 
-            battery_percentage_str = f"The battery percentage for {ROBOT_NAME} is currently {self.battery_states.charge_percentage}%.\n"
+            battery_percentage_str = f"The battery percentage for {self.robot_name} is currently {self.battery_states.charge_percentage}%.\n"
             remaining_runtime_str = f"The remaining runtime is estimated to {self.battery_states.estimated_runtime}.\n"
             temperatures_str = (
                 f"The temperature readings are {self.battery_states.temperatures}.\n"
@@ -531,9 +440,6 @@ class SpotController(Node):
             status_str = (
                 f"The current status of the battery is {self.battery_states.status}.\n"
             )
-
-            time_str = datetime.now().strftime("%H:%M:%S:%f")
-            print(Fore.CYAN + f"[{time_str}] End battery states time..")
 
             return (
                 battery_percentage_str
@@ -543,23 +449,9 @@ class SpotController(Node):
             )
 
         else:
-            time_str = datetime.now().strftime("%H:%M:%S:%f")
-            print(Fore.CYAN + f"[{time_str}] End battery states time..")
-            return f"Battery percentage for {ROBOT_NAME} has not been reported yet or is unavailable."
+            return f"Battery percentage for {self.robot_name} has not been reported yet or is unavailable."
 
     def get_mobility_metrics(self) -> str:
-        """
-        Reports mobility statistics of the robot dog.
-        The information that you can get through this tool include:
-        * The total distance traveled by the robot,
-        * The total time the robot has been moving,
-        * The current physical posture of the robot, namely whether or not the robot dog is standing/sitting and the desired body position and orientation (pose) relative to the world.
-        * The robot’s current motion configuration, namely whether or not the robot is currently moving, a locomotion behavior hint (e.g. walking, stair climbing), and whether the robot is actively in a stair-climbing mode.
-        """
-
-        time_str = datetime.now().strftime("%H:%M:%S:%f")
-        print(Fore.CYAN + f"[{time_str}] Start metrics time..")
-
         total_distance_str: str = "unknown"
         total_time_moving: str = "unknown"
         physical_posture: str = "unknown"
@@ -569,14 +461,14 @@ class SpotController(Node):
             total_distance_str = self.change_unknown_to_empty(total_distance_str)
             total_time_moving = self.change_unknown_to_empty(total_time_moving)
 
-            total_distance_str += f"The total distance traveled by {ROBOT_NAME} is {self.metrics.distance:.2f} meters.\n"
-            total_time_moving += f"The total time {ROBOT_NAME} has been moving is {self.metrics.time_moving.sec} seconds.\n"
+            total_distance_str += f"The total distance traveled by {self.robot_name} is {self.metrics.distance:.2f} meters.\n"
+            total_time_moving += f"The total time {self.robot_name} has been moving is {self.metrics.time_moving.sec} seconds.\n"
 
         if self.feedback is not None:
             physical_posture = self.change_unknown_to_empty(physical_posture)
             motion_configuration = self.change_unknown_to_empty(motion_configuration)
 
-            physical_posture += f"The current physical posture of {ROBOT_NAME} is :\nStanding : {self.feedback.standing}\nSitting:{self.feedback.sitting}.\n"
+            physical_posture += f"The current physical posture of {self.robot_name} is :\nStanding : {self.feedback.standing}\nSitting:{self.feedback.sitting}.\n"
             motion_configuration += f"The motion configuration of the robot is:\nMoving : {self.feedback.moving}\n"
 
         if self.mobility_params is not None:
@@ -585,9 +477,6 @@ class SpotController(Node):
 
             physical_posture += f"The estimated body position and orientation (pose) relative to the world is {self.mobility_params.body_control}\n"
             motion_configuration += f"Locomotion behavior hint: {self.mobility_params.locomotion_hint}\nStair climbing: {self.mobility_params.stair_hint}.\n"
-
-        time_str = datetime.now().strftime("%H:%M:%S:%f")
-        print(Fore.CYAN + f"[{time_str}] End metrics time..")
 
         return (
             "Total distance result: "
@@ -609,32 +498,14 @@ class SpotController(Node):
           * the estimated runtime remaining for locomotion.
           * the accumulated electrical power usage.
         """
-        time_str = datetime.now().strftime("%H:%M:%S:%f")
-        print(Fore.CYAN + f"[{time_str}] Start power state time..")
-
-        time_str = datetime.now().strftime("%H:%M:%S:%f")
-        print(Fore.CYAN + f"[{time_str}] End power state time..")
         return "Tool to be implemented."
 
     def get_info_spot(self) -> str:
         """Gets the robot dog (spot) general information, such as its specie, version and nickname"""
-
-        time_str = datetime.now().strftime("%H:%M:%S:%f")
-        print(Fore.CYAN + f"[{time_str}] Start info spot time..")
-
-        time_str = datetime.now().strftime("%H:%M:%S:%f")
-        print(Fore.CYAN + f"[{time_str}] End inof spot time..")
-
         return "Tool to be implemented."
 
     def get_odometry(self) -> str:
         """Represents the robot's current estimated velocity, both linear and angular, in three dimensions."""
-
-        time_str = datetime.now().strftime("%H:%M:%S:%f")
-        print(Fore.CYAN + f"[{time_str}] Start odometry time..")
-
-        time_str = datetime.now().strftime("%H:%M:%S:%f")
-        print(Fore.CYAN + f"[{time_str}] End odometry time..")
         return "Tool to be implemented."
 
     def get_faults(self) -> str:
@@ -645,30 +516,9 @@ class SpotController(Node):
         * System Fault State:
             Reports both active and historical faults detected in the robot’s internal systems.
         """
-        time_str = datetime.now().strftime("%H:%M:%S:%f")
-        print(Fore.CYAN + f"[{time_str}] Start faults time..")
-
-        time_str = datetime.now().strftime("%H:%M:%S:%f")
-        print(Fore.CYAN + f"[{time_str}] End faults time..")
         return "Tool to be implemented."
 
     def get_wifi_connection_state(self) -> str:
-        """Conveys the current Wi-Fi connection status of the robot.
-        The information that you can get through this tool include:
-        * the current operating mode of the wireless interface
-        * the name (SSID) of the currently connected network if applicable
-        The operating mode  is a number with the following meaning:
-            - 0 -> MODE UNKNOWN
-            - 1 -> MODE ACCESS POINT
-            - 2 -> MODE CLIENT
-
-        Please make sure to communicate the meanings (unknown, access point, client) to the user, as they might not know what a raw number  mean.
-
-        """
-
-        time_str = datetime.now().strftime("%H:%M:%S:%f")
-        print(Fore.CYAN + f"[{time_str}] Start wifi time..")
-
         if self.wifi_state is not None:
             mode_str = (
                 "unknown"
@@ -680,41 +530,20 @@ class SpotController(Node):
                 if self.wifi_state.ssid
                 else "unknown"
             )
-
-            time_str = datetime.now().strftime("%H:%M:%S:%f")
-            print(Fore.CYAN + f"[{time_str}] End wifi time..")
-
             return f"Wi-Fi Mode: {mode_str}.\n SSID: {ssid_str}."
         else:
-
-            time_str = datetime.now().strftime("%H:%M:%S:%f")
-            print(Fore.CYAN + f"[{time_str}] End wifi time..")
-
             return "Wi-Fi state has not been reported yet or is unavailable."
 
     def get_general_status(self) -> str:
-        """Gets the current robot dog's latest reported state and general information.
-        The status includes battery information, the robot's posture (standing, sitting, moving), wifi connection status, and the whether or not tracking is ongoing.
-        You should use this tool if the user simply asks for the status of the robot.
-        However, if they ask more details, please use the appropriate tools.
-        For more detailed information, use the appropriate tools, namely
-            * get_battery_status for battery information,
-            * get_mobility_metrics for mobility metrics,
-            * get_wifi_connection_state for the connection status
-        """
-
-        time_str = datetime.now().strftime("%H:%M:%S:%f")
-        print(Fore.CYAN + f"[{time_str}] Start general status time..")
-
         battery_percentage_str: str = "unknown"
         posture_str: str = "unknown"
         wifi_str: str = "unknown"
 
         if self.battery_states is not None:
-            battery_percentage_str = f"The battery percentage for {ROBOT_NAME} is currently {self.battery_states.charge_percentage}%.\n"
+            battery_percentage_str = f"The battery percentage for {self.robot_name} is currently {self.battery_states.charge_percentage}%.\n"
 
         if self.feedback is not None:
-            posture_str = f"The current physical posture of {ROBOT_NAME} is :\nStanding : {self.feedback.standing}\nSitting:{self.feedback.sitting}.\nThe motion configuration of the robot is:\nMoving : {self.feedback.moving}\n"
+            posture_str = f"The current physical posture of {self.robot_name} is :\nStanding : {self.feedback.standing}\nSitting:{self.feedback.sitting}.\nThe motion configuration of the robot is:\nMoving : {self.feedback.moving}\n"
 
         if self.wifi_state is not None:
             wifi_str = (
@@ -725,13 +554,10 @@ class SpotController(Node):
             )
 
         tracking_info_str = (
-            f" Actively tracking: '{current_tracking_object}'."
-            if current_tracking_object
+            f" Actively tracking: '{self.current_tracking_object}'."
+            if self.current_tracking_object
             else " Not currently tracking."
         )
-
-        time_str = datetime.now().strftime("%H:%M:%S:%f")
-        print(Fore.CYAN + f"[{time_str}] End general status time..")
 
         return (
             "Battery info :"
@@ -745,10 +571,6 @@ class SpotController(Node):
         )
 
     def sit(self) -> str:
-        """Command the robot dog to sit and transition from a standing position (on 4 legs) to a sitting position.
-        It cannot be used if the robot is already on the sitting. To know whether or not the robot is sitting, first get the status of the robot.
-        """
-
         # time stamp
         time_str = datetime.now().strftime("%H:%M:%S:%f")
         print(Fore.CYAN + f"[{time_str}] Start sit time..")
@@ -766,28 +588,13 @@ class SpotController(Node):
                 time_str = datetime.now().strftime("%H:%M:%S:%f")
                 print(Fore.CYAN + f"[{time_str}] End sit time..")
 
-                return f"{ROBOT_NAME} is sitting."
-
+                return f"{self.robot_name} is sitting."
             except Exception as e:
-                # time stamp
-                time_str = datetime.now().strftime("%H:%M:%S:%f")
-                print(Fore.CYAN + f"[{time_str}] End sit time..")
-
-                return f"Failed to sit {ROBOT_NAME}: {e}"
+                return f"Failed to sit {self.robot_name}: {e}"
         else:
-            # time stamp
-            time_str = datetime.now().strftime("%H:%M:%S:%f")
-            print(Fore.CYAN + f"[{time_str}] End sit time..")
-
             return f"Robot state (sitting/standing) is not yet known. Cannot sit."
 
     def stand(self) -> str:
-        """Command the robot to stand and transition from a sitting position to a standing position.
-        It cannot be used if the robot is already standing.  To know whether or not the robot is sitting, first get the status of the robot.
-        Do not attempt to stand the robot if the battery level is below 20% or if the robot is standing. In such cases, return an appropriate message.
-        To have the battery level, use the get_battery_status() tool.
-        """
-
         # time stamp
         time_str = datetime.now().strftime("%H:%M:%S:%f")
         print(Fore.CYAN + f"[{time_str}] Start stand tool time..")
@@ -804,44 +611,13 @@ class SpotController(Node):
                 time_str = datetime.now().strftime("%H:%M:%S:%f")
                 print(Fore.CYAN + f"[{time_str}] End stand tool time..")
 
-                return f"{ROBOT_NAME} is standing."
+                return f"{self.robot_name} is standing."
             except Exception as e:
-
-                time_str = datetime.now().strftime("%H:%M:%S:%f")
-                print(Fore.CYAN + f"[{time_str}] End stand tool time..")
-
-                return f"Failed to stand {ROBOT_NAME}: {e}"
+                return f"Failed to stand {self.robot_name}: {e}"
         else:
-            time_str = datetime.now().strftime("%H:%M:%S:%f")
-            print(Fore.CYAN + f"[{time_str}] End stand tool time..")
-
             return f"Robot state (sitting/standing) is not yet known. Cannot stand."
 
     def move(self, linear: List[float], angular: float, duration: int) -> str:
-        """
-        Move the robot with specified linear and angular velocities for a given duration.
-        This function controls movement along three axes (x, y, z) and rotation.
-
-        The coordinate system is as follows:
-        - x-axis: +x is forward, -x is backward.
-        - y-axis: +y is left, -y is right.
-        - z-axis: +z is up, -z is down.
-        - angular z-axis: +z is counter-clockwise turn (left), -z is clockwise turn (right).
-
-        To perform this movement, the robot dog must be in the standing.
-        The user may specify a speed (e.g., "velocity 1m/s", "go slowly"). If a speed is provided, use it to set the magnitude of the linear or angular velocity vector. If no speed is specified, use a default of 1 m/s or -1 m/s.
-        If the user does not specify a time, assume a default duration of 1 second.
-
-        :param linear: A list of 3 floats representing [x, y, z] velocity in m/s. This vector should be constructed based on the user's direction and specified speed.
-                    For example, if the user says "go right at 1.2 m/s", the vector should be [0.0, -1.2, 0.0].
-        :param angular: A float for z-axis angular velocity (rotation).
-        :param duration: Duration of the movement in seconds.
-
-        Do not attempt to move the robot if the battery level is below 20% or if the robot is sitting. In such cases, return an appropriate message.
-        To have the battery level and whether or not the robot is sitting, you have the , use the get_general_status() tool.
-        Remember, for the parameters, you have to provide floats, not integers!
-        """
-
         # time stamp
         time_str = datetime.now().strftime("%H:%M:%S:%f")
         print(Fore.CYAN + f"[{time_str}] Start motion time..")
@@ -869,7 +645,7 @@ class SpotController(Node):
                 t0 = self.get_clock().now().nanoseconds
 
                 while (self.get_clock().now().nanoseconds - t0) / 1e9 <= duration:
-                    # print((self.get_clock().now().nanoseconds - t0) / 1e9)
+                    print((self.get_clock().now().nanoseconds - t0) / 1e9)
 
                     self.commands_pub.publish(msg_twist)
 
@@ -880,35 +656,14 @@ class SpotController(Node):
                 time_str = datetime.now().strftime("%H:%M:%S:%f")
                 print(Fore.CYAN + f"[{time_str}] End motion time..")
 
-                return f"Moved {ROBOT_NAME} with linear={linear}, angular={angular} for {duration}s and then stopped."
+                return f"Moved {self.robot_name} with linear={linear}, angular={angular} for {duration}s and then stopped."
             except Exception as e:
-                # time stamp
-                time_str = datetime.now().strftime("%H:%M:%S:%f")
-                print(Fore.CYAN + f"[{time_str}] End motion time..")
-
-                print(f"Failed to move {ROBOT_NAME}: {e}")
-                return f"Failed to move {ROBOT_NAME}: {e}"
+                print(f"Failed to move {self.robot_name}: {e}")
+                return f"Failed to move {self.robot_name}: {e}"
         else:
-            # time stamp
-            time_str = datetime.now().strftime("%H:%M:%S:%f")
-            print(Fore.CYAN + f"[{time_str}] End motion time..")
             return f"Robot state (sitting/standing) is not yet known. Cannot move."
 
     def switch_mode(self, mode: str, object_name: Optional[str] = None) -> str:
-        """
-        Switches the control mode of the robot dog. Tell the user that he has to select the image window.
-        The LLM should request modes like 'keyboard', 'hand', 'tracking' or 'stop tracking'.
-        If the user selects 'keyboard', he has to know that to stand he has to use "t", to sit "g", to move the letters "a", "w", "d", "s". Still in keyboard mode
-        the user must know that for rotation he/she can use "left-arrow" and "right-arrow" to move left or right respectively, and for altitude, he/she can use "up-arrow" to move up and "down-arrow" to move down.
-        If the user selects 'hand', he has to know that he has to use the hands to control the robot dog, all the options are in the image window.
-        If the user selects 'tracking', tracking': Start tracking a person holding a specific object.
-        When using this mode, you must also provide the 'object_name' parameter.
-        Choose the object from this list: [backpack, umbrella, handbag, bottle, cup, fork, knife, spoon, bowl, banana, apple, cell phone, book].
-        If the user selects 'stop tracking': Stop the current tracking task.
-
-        :param mode: The desired control mode as a string (e.g., "keyboard", "hand", "tracking", "stop tracking").
-        :param object_name: The name of the object to track. Required only for 'tracking' mode.
-        """
         # time stamp
         time_str = datetime.now().strftime("%H:%M:%S:%f")
         print(Fore.CYAN + f"[{time_str}] Start switch mode time..")
@@ -934,10 +689,6 @@ class SpotController(Node):
 
             elif mode_requested == "tracking":
                 if not object_name:
-                    # time stamp
-                    time_str_end = datetime.now().strftime("%H:%M:%S:%f")
-                    print(Fore.CYAN + f"[{time_str_end}] End switch mode time..")
-
                     return "Error: To switch to tracking mode, you must specify an object_name."
                 msg_str.data = "t"
                 self.key_pressed_pub.publish(msg_str)
@@ -959,17 +710,23 @@ class SpotController(Node):
             return response
 
         except Exception as e:
-            # time stamp
-            time_str_end = datetime.now().strftime("%H:%M:%S:%f")
-            print(Fore.CYAN + f"[{time_str_end}] End switch mode time..")
-
             return (
-                f"Failed to switch mode for {ROBOT_NAME} to '{mode_requested}' "
+                f"Failed to switch mode for {self.robot_name} to '{mode_requested}' "
                 f"(as '{msg_str.data}'): {e}"
             )
 
     def start_object_tracking(self, object_name: str) -> str:
         """
+        Use this tool to start tracking a person holding a specific object. It will wait
+        up to 5 seconds for a person holding this object to be detected. The robot MUST be  standing.
+
+        First, you MUST choose the most similar object from this list of available options:
+        [backpack, umbrella, handbag, bottle, cup, fork, knife, spoon, bowl, banana, apple, cell phone, book]
+        Match the user's request to an object in the list. For example, if the user asks for a "phone", choose "cell phone".
+        If you cannot find a clear match, respond by saying "There are no similar objects to track."
+        :param object_name: str - The chosen object name from the list.
+
+
         Use this tool to start tracking a person holding a specific object. It will wait
         up to 5 seconds for a person holding this object to be detected. The robot MUST be  standing.
 
@@ -984,38 +741,32 @@ class SpotController(Node):
         time_str = datetime.now().strftime("%H:%M:%S:%f")
         print(Fore.CYAN + f"[{time_str}] Start start_tracking tool time..")
 
-        global current_tracking_object, tracking_confirmation_received, last_published_data
-
-        if current_tracking_object:
-            return f"Already tracking '{current_tracking_object}'. Please stop tracking first."
+        if self.current_tracking_object:
+            return f"Already tracking '{self.current_tracking_object}'. Please stop tracking first."
 
         # Reset the event and prepare for a new tracking task
-        tracking_confirmation_received.clear()
-        last_published_data = None
-        current_tracking_object = object_name
+        self.current_tracking_object = object_name
+        self.tracking_confirmation_received.clear()
 
-        node = globals().get("node")
-        if node:
-            node.get_logger().info(
-                f"Attempting to track a person with a '{current_tracking_object}'. Searching for 5 seconds..."
-            )
+        self.get_logger().info(
+            f"Attempting to track a person with a '{self.current_tracking_object}'. Searching for 5 seconds..."
+        )
 
-        success = tracking_confirmation_received.wait(timeout=5.0)
+        success = self.tracking_confirmation_received.wait(timeout=5.0)
 
         if success:
-            response = f"Successfully found and locked on to person with '{current_tracking_object}'. Published data: {last_published_data}"
-            if node:
-                node.get_logger().info(response)
+            response = f"Successfully found and locked on to person with '{self.current_tracking_object}'."
+
+            self.get_logger().info(response)
 
             time_str = time.strftime("%H:%M:%S")
             print(Fore.CYAN + f"[{time_str}] End tracking time..")
 
             return response
         else:
-            current_tracking_object = None
+            self.current_tracking_object = None
             response = f"Failed to find a person with a '{object_name}' within 5 seconds. Please ensure they and the object are visible."
-            if node:
-                node.get_logger().warn(response)
+            self.get_logger().warn(response)
 
             time_str = datetime.now().strftime("%H:%M:%S:%f")
             print(Fore.CYAN + f"[{time_str}] End start_tracking tool time..")
@@ -1028,12 +779,13 @@ class SpotController(Node):
         """
         Stops tracking the current object and clears the tracking target.
         An explicit "stop_tracking" message is sent to the tracking system.
+
+        Stops tracking the current object and clears the tracking target.
+        An explicit "stop_tracking" message is sent to the tracking system.
         """
 
         time_str = datetime.now().strftime("%H:%M:%S:%f")
         print(Fore.CYAN + f"[{time_str}] Start stop_tracking tool time..")
-
-        global current_tracking_object
 
         try:
             stop_tracking_message = json.dumps(
@@ -1041,373 +793,219 @@ class SpotController(Node):
             )
             self.tracking_signal_pub.publish(String(data=stop_tracking_message))
         except Exception as e:
-            return f"Failed to publish stop tracking command for {ROBOT_NAME}: {e}"
-        previous_object = current_tracking_object
-        current_tracking_object = None
+            return f"Failed to publish stop tracking command for {self.robot_name}: {e}"
+        previous_object = self.current_tracking_object
+        self.current_tracking_object = None
 
         time_str = datetime.now().strftime("%H:%M:%S:%f")
         print(Fore.CYAN + f"[{time_str}] End stop_tracking tool time..")
 
         if previous_object:
-            return f"{ROBOT_NAME} has stopped tracking '{previous_object}'. Stop command was successfully published."
+            return f"{self.robot_name} has stopped tracking '{previous_object}'. Stop command was successfully published."
         else:
-            return f"A stop command was sent to {ROBOT_NAME} to ensure tracking is disabled."
-
-
-# def print_response(query: str):
-#     response = rosa.invoke(query)
-#     # ... code to display response ...
-################################################### Tools #############################################################
-
-
-async def submit(query: str, rosa):
-    return await stream_response(query, rosa)
-
-
-async def stream_response(query: str, rosa):
-    print(Fore.BLUE + Style.BRIGHT + f"\n👤 User: {query}\n")
-
-    response = ""
-
-    async for event in rosa.astream(query):
-        if event["type"] == "token":
-            print(Fore.GREEN + event["content"], end="", flush=True)
-            response = response + event["content"]
-        elif event["type"] == "tool_start":
-            print(Fore.YELLOW + f"\n🛠️ Starting tool: {event['name']}")
-        elif event["type"] == "tool_end":
-            print(Fore.YELLOW + f"\n✅ Finished tool: {event['name']}")
-            await asyncio.sleep(1)
-        elif event["type"] == "final":
-            pass
-            # print(Fore.CYAN + Style.BRIGHT + f"\n📤 Final output: {event['content']}")
-        elif event["type"] == "error":
-            print(Fore.RED + f"\n❌ Error: {event['content']}")
-
-    return response
-
-
-# Function to run the ROSA agent
-async def run(node, rosa):
-    # Print the time when the session starts
-    start_time_str = time.strftime("%Y-%m-%d %H:%M:%S")
-    print(
-        Fore.CYAN + Style.BRIGHT + f"\n--- ROSA Session Started at {start_time_str} ---"
-    )
-
-    query = None
-
-    while True:
-
-        # # Get the user's command
-        # if node.user_query != query:
-        #     query = node.user_query
-
-        #     # --- 1. Print the time BEFORE the prompt ---
-        #     prompt_time_str = time.strftime("%H:%M:%S")
-        #     # The '\n' adds a space before the new prompt, 'end=""' keeps the cursor on the same line
-        #     print(Fore.CYAN + f"\n[{prompt_time_str}] ", end="")
-
-        #     query = input("Enter your prompt (or 'exit' to quit): ")
-
-        #     if query.lower() == "exit":
-        #         log.info("--- Session Ended ---")
-        #         break
-
-        #     # Log the prompt to the file
-        #     log.info(f"USER_PROMPT: {query}")
-
-        #     # Start processing the command
-        #     processing_time_str = time.strftime("%H:%M:%S")
-        #     print(Fore.CYAN + f"[{processing_time_str}] Processing command...")
-
-        #     response = await submit(query, rosa)
-
-        #     responseMsg = String()
-        #     responseMsg.data = response
-        #     node.llm_response_pub.publish(responseMsg)
-
-        #     processing_time_str = time.strftime("%H:%M:%S")
-        #     print(Fore.CYAN + f"[{processing_time_str}] Response received...")
-
-        # --- 1. Print the time BEFORE the prompt ---
-        prompt_time_str = time.strftime("%H:%M:%S")
-        # The '\n' adds a space before the new prompt, 'end=""' keeps the cursor on the same line
-        print(Fore.CYAN + f"\n[{prompt_time_str}] ", end="")
-
-        query = input("Enter your prompt (or 'exit' to quit): ")
-
-        if query.lower() == "exit":
-            log.info("--- Session Ended ---")
-            break
-
-        # Log the prompt to the file
-        log.info(f"USER_PROMPT: {query}")
-
-        # Start processing the command
-        processing_time_str = time.strftime("%H:%M:%S")
-        print(Fore.CYAN + f"[{processing_time_str}] Processing command...")
-
-        response = await submit(query, rosa)
-
-        responseMsg = String()
-        responseMsg.data = response
-        node.llm_response_pub.publish(responseMsg)
-
-        processing_time_str = time.strftime("%H:%M:%S")
-        print(Fore.CYAN + f"[{processing_time_str}] Response received...")
-
-
-# asyncio.run(run())
-
-
-# ROS init and run
-def main(args=None):
-    rclpy.init(args=args)
-    node = SpotController()
-    # node.get_logger().set_level(rclpy.logging.LoggingSeverity.DEBUG)
-
-    # Use executor in a separate thread
-    executor = MultiThreadedExecutor()
-    executor.add_node(node)
-
-    def spin():
-        executor.spin()
-
-    # Start the ROS spinning in a background thread
-    spin_thread = threading.Thread(target=spin, daemon=True)
-    spin_thread.start()
-
-    prompts = RobotSystemPrompts(
-        embodiment_and_persona="You are a robotic agent managing a robot dog. The robot dog you are operating is Boston dynamics' Spot",
-        about_your_capabilities="You capabilities are limited to the available tools. Anything that is asked to you and not provided by a tool is beyond your capabilities",
-        critical_instructions="Always use the corresponding tool if you can. If the user ask you to perform an action requiring to move the robot, always use the mode tool. Sam for all other tools: if the user ask for an information/action requiring to use a tool, always use the relevant tool. Also, tell the user what you are trying to do.",
-    )
-
-    ################################## tools ########################################################
-
-    @tool
-    def get_battery_status(node=node):
-        """Provides detailed information about each battery installed in the robot.
-        For instance, it includes :
-        * the percentage of charge remaining,
-        * the estimated remaining runtime,
-        * the temperature readings.
-        * a status indicator to flag battery health or errors.
-
-        The status indicator is a number with the following meaning:
-            - 0 -> STATUS UNKNOWN
-            - 1 -> STATUS MISSING
-            - 2 -> STATUS CHARGING
-            - 3 -> STATUS DISCHARGING
-            - 4 -> STATUS BOOTING
-        Please make sure to communicate the meanings to the user, as they might not know what a raw number for the battery status mean.
-
-        This should be called when necessary to guarantee accurate battery level information.
-        """
-        return node.get_battery_status()
-
-    @tool
-    def get_mobility_metrics(node=node):
-        """
-        Reports mobility statistics of the robot dog.
-        The information that you can get through this tool include:
-        * The total distance traveled by the robot,
-        * The total time the robot has been moving,
-        * The current physical posture of the robot, namely whether or not the robot dog is standing/sitting and the desired body position and orientation (pose) relative to the world.
-        * The robot’s current motion configuration, namely whether or not the robot is currently moving, a locomotion behavior hint (e.g. walking, stair climbing), and whether the robot is actively in a stair-climbing mode.
-        """
-        return node.get_mobility_metrics()
-
-    @tool
-    def get_power_state(node=node):
-        """
-        Provides the current power supply conditions.
-        The information that you can get through this tool include:
-        * the state of the motor and shore power systems,
-        * the charge level specific to locomotion systems,
-        * the estimated runtime remaining for locomotion.
-        * the accumulated electrical power usage.
-        """
-        return node.get_power_state()
-
-    @tool
-    def get_info_spot(node=node):
-        """Gets the robot dog (spot) general information, such as its specie, version and nickname"""
-        return node.get_info_spot()
-
-    @tool
-    def get_odometry(node=node):
-        """Represents the robot's current estimated velocity, both linear and angular, in three dimensions."""
-        return node.get_odometry()
-
-    @tool
-    def get_faults(node=node):
-        """
-        Gets faults reports of the robot dog. There are two faults types:
-        * Behavior Fault State:
-            Lists any active behavior-related faults.
-        * System Fault State:
-            Reports both active and historical faults detected in the robot’s internal systems.
-        """
-        return node.get_faults()
-
-    @tool
-    def get_wifi_connection_state(node=node):
-        """Conveys the current Wi-Fi connection status of the robot.
-        The information that you can get through this tool include:
-        * the current operating mode of the wireless interface
-        * the name (SSID) of the currently connected network if applicable
-        The operating mode  is a number with the following meaning:
-            - 0 -> MODE UNKNOWN
-            - 1 -> MODE ACCESS POINT
-            - 2 -> MODE CLIENT
-
-        Please make sure to communicate the meanings (unknown, access point, client) to the user, as they might not know what a raw number  mean.
-
-        """
-        return node.get_wifi_connection_state()
-
-    @tool
-    def get_general_status(node=node):
-        """Gets the current robot dog's latest reported state and general information.
-        The status includes battery information, the robot's posture (standing, sitting, moving), wifi connection status, and the whether or not tracking is ongoing.
-        You should use this tool if the user simply asks for the status of the robot.
-        However, if they ask more details, please use the appropriate tools.
-        For more detailed information, use the appropriate tools, namely
-            * get_battery_status for battery information,
-            * get_mobility_metrics for mobility metrics,
-            * get_wifi_connection_state for the connection status
-        """
-        return node.get_general_status()
-
-    @tool
-    def sit(node=node):
-        """Command the robot dog to sit and transition from a standing position (on 4 legs) to a sitting position.
-        It cannot be used if the robot is already on the sitting. To know whether or not the robot is sitting, first get the status of the robot.
-        """
-        return node.sit()
-
-    @tool
-    def stand(node=node):
-        """Command the robot to stand and transition from a sitting position to a standing position.
-        It cannot be used if the robot is already standing.  To know whether or not the robot is sitting, first get the status of the robot.
-        Do not attempt to stand the robot if the battery level is below 20% or if the robot is standing. In such cases, return an appropriate message.
-        To have the battery level, use the get_battery_status() tool.
-        """
-        return node.stand()
-
-    @tool
-    def move(linear, angular, duration, node=node):
-        """
-        Move the robot with specified linear and angular velocities for a given duration.
-        This function controls movement along three axes (x, y, z) and rotation.
-
-        The coordinate system is as follows:
-        - x-axis: +x is forward, -x is backward.
-        - y-axis: +y is left, -y is right.
-        - z-axis: +z is up, -z is down.
-        - angular z-axis: +z is counter-clockwise turn (left), -z is clockwise turn (right).
-
-        To perform this movement, the robot dog must be in the standing.
-        The user may specify a speed (e.g., "velocity 1m/s", "go slowly"). If a speed is provided, use it to set the magnitude of the linear or angular velocity vector. If no speed is specified, use a default of 1 m/s or -1 m/s.
-        If the user does not specify a time, assume a default duration of 1 second.
-
-        :param linear: A list of 3 floats representing [x, y, z] velocity in m/s. This vector should be constructed based on the user's direction and specified speed.
-                    For example, if the user says "go right at 1.2 m/s", the vector should be [0.0, -1.2, 0.0].
-        :param angular: A float for z-axis angular velocity (rotation).
-        :param duration: Duration of the movement in seconds.
-
-        Do not attempt to move the robot if the battery level is below 20% or if the robot is sitting. In such cases, return an appropriate message.
-        To have the battery level and whether or not the robot is sitting, you have the , use the get_general_status() tool.
-        Remember, for the parameters, you have to provide floats, not integers!
-        """
-        return node.move(linear, angular, duration)
-
-    @tool
-    def switch_mode(mode, object_name, node=node):
-        """
-        Switches the control mode of the robot dog. Tell the user that he has to select the image window.
-        The LLM should request modes like 'keyboard', 'hand', 'tracking' or 'stop tracking'.
-        If the user selects 'keyboard', he has to know that to stand he has to use "t", to sit "g", to move the letters "a", "w", "d", "s". Still in keyboard mode
-        the user must know that for rotation he/she can use "left-arrow" and "right-arrow" to move left or right respectively, and for altitude, he/she can use "up-arrow" to move up and "down-arrow" to move down.
-        If the user selects 'hand', he has to know that he has to use the hands to control the robot dog, all the options are in the image window.
-        If the user selects 'tracking', tracking': Start tracking a person holding a specific object.
-        When using this mode, you must also provide the 'object_name' parameter.
-        Choose the object from this list: [backpack, umbrella, handbag, bottle, cup, fork, knife, spoon, bowl, banana, apple, cell phone, book].
-        If the user selects 'stop tracking': Stop the current tracking task.
-
-        :param mode: The desired control mode as a string (e.g., "keyboard", "hand", "tracking", "stop tracking").
-        :param object_name: The name of the object to track. Required only for 'tracking' mode.
-        """
-        return node.switch_mode(mode, object_name)
-
-    @tool
-    def start_object_tracking(object_name, node=node):
-        """
-        Use this tool to start tracking a person holding a specific object. It will wait
-        up to 5 seconds for a person holding this object to be detected. The robot MUST be  standing.
-
-        First, you MUST choose the most similar object from this list of available options:
-        [backpack, umbrella, handbag, bottle, cup, fork, knife, spoon, bowl, banana, apple, cell phone, book]
-        Match the user's request to an object in the list. For example, if the user asks for a "phone", choose "cell phone".
-        If you cannot find a clear match, respond by saying "There are no similar objects to track."
-        :param object_name: str - The chosen object name from the list.
-        """
-        return node.start_object_tracking(object_name)
-
-    @tool
-    def stop_object_tracking(node=node):
-        """
-        Stops tracking the current object and clears the tracking target.
-        An explicit "stop_tracking" message is sent to the tracking system.
-        """
-        return node.stop_object_tracking()
-
-    tools = [
-        get_battery_status,
-        get_mobility_metrics,
-        get_info_spot,
-        get_faults,
-        get_odometry,
-        get_general_status,
-        get_power_state,
-        stand,
-        sit,
-        start_object_tracking,
-        stop_object_tracking,
-        move,
-        switch_mode,
-        get_wifi_connection_state,
-    ]
-
-    llm_model = ChatOpenAI(
-        model="gpt-4-turbo",
-        temperature=0,
-        timeout=None,
-        max_retries=2,
-    )
-
-    # Pass the LLM to ROSA
-    rosa = ROSA(
-        ros_version=2, llm=llm_model, streaming=True, tools=tools, prompts=prompts
-    )
-
-    try:
-        asyncio.run(run(node, rosa))
-    except KeyboardInterrupt:
-        pass
-    finally:
-        executor.shutdown()
-        node.destroy_node()
-        rclpy.shutdown()
-
-
-if __name__ == "__main__":
-    main()
-
-
-#     def tracking_signal_print_callback(self, msg: String):
-#         self.get_logger().info("--- FINAL TRACKING SIGNAL (Output) ---")
-#         self.get_logger().info(f"{msg.data}")
-#         self.get_logger().info("------------------------------------")
+            return f"A stop command was sent to {self.robot_name} to ensure tracking is disabled."
+
+    ############################################# Getters #########################################################################
+    def get_prompts(self):
+        prompts = RobotSystemPrompts(
+            embodiment_and_persona="You are a robotic agent managing a robot dog. The robot dog you are operating is Boston dynamics' Spot",
+            about_your_capabilities="You capabilities are limited to the available tools. Anything that is asked to you and not provided by a tool is beyond your capabilities",
+            critical_instructions="Always use the corresponding tool if you can. If the user ask you to perform an action requiring to move the robot, always use the mode tool. Sam for all other tools: if the user ask for an information/action requiring to use a tool, always use the relevant tool. Also, tell the user what you are trying to do.",
+        )
+        return prompts
+
+    def get_tools(self):
+
+        @tool
+        def get_battery_status():
+            """Provides detailed information about each battery installed in the robot.
+            For instance, it includes :
+            * the percentage of charge remaining,
+            * the estimated remaining runtime,
+            * the temperature readings.
+            * a status indicator to flag battery health or errors.
+
+            The status indicator is a number with the following meaning:
+                - 0 -> STATUS UNKNOWN
+                - 1 -> STATUS MISSING
+                - 2 -> STATUS CHARGING
+                - 3 -> STATUS DISCHARGING
+                - 4 -> STATUS BOOTING
+            Please make sure to communicate the meanings to the user, as they might not know what a raw number for the battery status mean.
+
+            This should be called when necessary to guarantee accurate battery level information.
+            """
+            return self.get_battery_status()
+
+        @tool
+        def get_mobility_metrics():
+            """
+            Reports mobility statistics of the robot dog.
+            The information that you can get through this tool include:
+            * The total distance traveled by the robot,
+            * The total time the robot has been moving,
+            * The current physical posture of the robot, namely whether or not the robot dog is standing/sitting and the desired body position and orientation (pose) relative to the world.
+            * The robot’s current motion configuration, namely whether or not the robot is currently moving, a locomotion behavior hint (e.g. walking, stair climbing), and whether the robot is actively in a stair-climbing mode.
+            """
+            return self.get_mobility_metrics()
+
+        @tool
+        def get_power_state():
+            """
+            Provides the current power supply conditions.
+            The information that you can get through this tool include:
+            * the state of the motor and shore power systems,
+            * the charge level specific to locomotion systems,
+            * the estimated runtime remaining for locomotion.
+            * the accumulated electrical power usage.
+            """
+            return self.get_power_state()
+
+        @tool
+        def get_info_spot():
+            """Gets the robot dog (spot) general information, such as its specie, version and nickname"""
+            return self.get_info_spot()
+
+        @tool
+        def get_odometry():
+            """Represents the robot's current estimated velocity, both linear and angular, in three dimensions."""
+            return self.get_odometry()
+
+        @tool
+        def get_faults():
+            """
+            Gets faults reports of the robot dog. There are two faults types:
+            * Behavior Fault State:
+                Lists any active behavior-related faults.
+            * System Fault State:
+                Reports both active and historical faults detected in the robot’s internal systems.
+            """
+            return self.get_faults()
+
+        @tool
+        def get_wifi_connection_state():
+            """Conveys the current Wi-Fi connection status of the robot.
+            The information that you can get through this tool include:
+            * the current operating mode of the wireless interface
+            * the name (SSID) of the currently connected network if applicable
+            The operating mode  is a number with the following meaning:
+                - 0 -> MODE UNKNOWN
+                - 1 -> MODE ACCESS POINT
+                - 2 -> MODE CLIENT
+
+            Please make sure to communicate the meanings (unknown, access point, client) to the user, as they might not know what a raw number  mean.
+
+            """
+            return self.get_wifi_connection_state()
+
+        @tool
+        def get_general_status():
+            """Gets the current robot dog's latest reported state and general information.
+            The status includes battery information, the robot's posture (standing, sitting, moving), wifi connection status, and the whether or not tracking is ongoing.
+            You should use this tool if the user simply asks for the status of the robot.
+            However, if they ask more details, please use the appropriate tools.
+            For more detailed information, use the appropriate tools, namely
+                * get_battery_status for battery information,
+                * get_mobility_metrics for mobility metrics,
+                * get_wifi_connection_state for the connection status
+            """
+            return self.get_general_status()
+
+        @tool
+        def sit():
+            """Command the robot dog to sit and transition from a standing position (on 4 legs) to a sitting position.
+            It cannot be used if the robot is already on the sitting. To know whether or not the robot is sitting, first get the status of the robot.
+            """
+            return self.sit()
+
+        @tool
+        def stand():
+            """Command the robot to stand and transition from a sitting position to a standing position.
+            It cannot be used if the robot is already standing.  To know whether or not the robot is sitting, first get the status of the robot.
+            Do not attempt to stand the robot if the battery level is below 20% or if the robot is standing. In such cases, return an appropriate message.
+            To have the battery level, use the get_battery_status() tool.
+            """
+            return self.stand()
+
+        @tool
+        def move(linear, angular, duration):
+            """
+            Move the robot with specified linear and angular velocities for a given duration.
+            This function controls movement along three axes (x, y, z) and rotation.
+
+            The coordinate system is as follows:
+            - x-axis: +x is forward, -x is backward.
+            - y-axis: +y is left, -y is right.
+            - z-axis: +z is up, -z is down.
+            - angular z-axis: +z is counter-clockwise turn (left), -z is clockwise turn (right).
+
+            To perform this movement, the robot dog must be in the standing.
+            The user may specify a speed (e.g., "velocity 1m/s", "go slowly"). If a speed is provided, use it to set the magnitude of the linear or angular velocity vector. If no speed is specified, use a default of 1 m/s or -1 m/s.
+            If the user does not specify a time, assume a default duration of 1 second.
+
+            :param linear: A list of 3 floats representing [x, y, z] velocity in m/s. This vector should be constructed based on the user's direction and specified speed.
+                        For example, if the user says "go right at 1.2 m/s", the vector should be [0.0, -1.2, 0.0].
+            :param angular: A float for z-axis angular velocity (rotation).
+            :param duration: Duration of the movement in seconds.
+
+            Do not attempt to move the robot if the battery level is below 20% or if the robot is sitting. In such cases, return an appropriate message.
+            To have the battery level and whether or not the robot is sitting, you have the , use the get_general_status() tool.
+            """
+            return self.move(linear, angular, duration)
+
+        @tool
+        def switch_mode(mode, object_name):
+            """
+            Switches the control mode of the robot dog. Tell the user that he has to select the image window.
+            The LLM should request modes like 'keyboard', 'hand', 'tracking' or 'stop tracking'.
+            If the user selects 'keyboard', he has to know that to stand he has to use "t", to sit "g", to move the letters "a", "w", "d", "s". Still in keyboard mode
+            the user must know that for rotation he/she can use "left-arrow" and "right-arrow" to move left or right respectively, and for altitude, he/she can use "up-arrow" to move up and "down-arrow" to move down.
+            If the user selects 'hand', he has to know that he has to use the hands to control the robot dog, all the options are in the image window.
+            If the user selects 'tracking', tracking': Start tracking a person holding a specific object.
+            When using this mode, you must also provide the 'object_name' parameter.
+            Choose the object from this list: [backpack, umbrella, handbag, bottle, cup, fork, knife, spoon, bowl, banana, apple, cell phone, book].
+            If the user selects 'stop tracking': Stop the current tracking task.
+
+            :param mode: The desired control mode as a string (e.g., "keyboard", "hand", "tracking", "stop tracking").
+            :param object_name: The name of the object to track. Required only for 'tracking' mode.
+            """
+            return self.switch_mode(mode, object_name)
+
+        @tool
+        def start_object_tracking(object_name):
+            """
+            Use this tool to start tracking a person holding a specific object. It will wait
+            up to 5 seconds for a person holding this object to be detected. The robot MUST be  standing.
+
+            First, you MUST choose the most similar object from this list of available options:
+            [backpack, umbrella, handbag, bottle, cup, fork, knife, spoon, bowl, banana, apple, cell phone, book]
+            Match the user's request to an object in the list. For example, if the user asks for a "phone", choose "cell phone".
+            If you cannot find a clear match, respond by saying "There are no similar objects to track."
+            :param object_name: str - The chosen object name from the list.
+            """
+            return self.start_object_tracking(object_name)
+
+        @tool
+        def stop_object_tracking():
+            """
+            Stops tracking the current object and clears the tracking target.
+            An explicit "stop_tracking" message is sent to the tracking system.
+            """
+            return self.stop_object_tracking()
+
+        return [
+            get_battery_status,
+            get_mobility_metrics,
+            get_info_spot,
+            get_faults,
+            get_odometry,
+            get_general_status,
+            get_power_state,
+            stand,
+            sit,
+            move,
+            switch_mode,
+            start_object_tracking,
+            stop_object_tracking,
+            get_wifi_connection_state,
+        ]
